Move more tests using Cell to unit tests
