+++ library/core/tests/option.rs
@@ -372,3 +372,34 @@ fn option_const() {
     const IS_NONE: bool = OPTION.is_none();
     assert!(!IS_NONE);
 }
+
+#[test]
+fn test_unwrap_drop() {
+    use std::cell::Cell;
+
+    struct Dtor<'a> {
+        x: &'a Cell<isize>,
+    }
+
+    impl<'a> std::ops::Drop for Dtor<'a> {
+        fn drop(&mut self) {
+            self.x.set(self.x.get() - 1);
+        }
+    }
+
+    fn unwrap<T>(o: Option<T>) -> T {
+        match o {
+            Some(v) => v,
+            None => panic!(),
+        }
+    }
+
+    let x = &Cell::new(1);
+
+    {
+        let b = Some(Dtor { x });
+        let _c = unwrap(b);
+    }
+
+    assert_eq!(x.get(), 0);
+}
